+++ Hacker-Stories/src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './InputWithLabel';
+export * from './List';
+export * from './SearchForm';
