+++ Hacker-Stories/src/storiesReducer.ts
@@ -1,0 +1,92 @@
+interface Story {
+  title: string;
+  url: string;
+  author: string;
+  num_comments: number;
+  points: number;
+  objectID: string;
+}
+
+enum Action {
+  SET_STORIES = "SET_STORIES",
+  REMOVE_STORY = "REMOVE_STORY",
+  STORIES_FETCH_INIT = "STORIES_FETCH_INIT",
+  STORIES_FETCH_SUCCESS = "STORIES_FETCH_SUCCESS",
+  STORIES_FETCH_FAILURE = "STORIES_FETCH_FAILURE",
+}
+
+type StoriesInitAction = {
+  type: Action.STORIES_FETCH_INIT;
+};
+
+type StoriesSuccessAction = {
+  type: Action.STORIES_FETCH_SUCCESS;
+  payload: Array<Story>;
+};
+
+type StoriesFailureAction = {
+  type: Action.STORIES_FETCH_FAILURE;
+};
+
+type StoriesSetAction = {
+  type: Action.SET_STORIES;
+  payload: Array<Story>;
+};
+
+type StoriesRemoveAction = {
+  type: Action.REMOVE_STORY;
+  payload: Story;
+};
+
+type StoriesAction =
+  | StoriesSetAction
+  | StoriesRemoveAction
+  | StoriesInitAction
+  | StoriesSuccessAction
+  | StoriesFailureAction;
+
+interface StoryState {
+  data: Array<Story>;
+  isLoading: boolean;
+  isError: boolean;
+}
+
+type StoryReducer = (state: StoryState, action: StoriesAction) => StoryState;
+
+const storiesReducer: StoryReducer = (
+  state: StoryState,
+  action: StoriesAction
+) => {
+  switch (action.type) {
+    case Action.STORIES_FETCH_INIT:
+      return {
+        ...state,
+        isLoading: true,
+        isError: false,
+      };
+    case Action.STORIES_FETCH_SUCCESS:
+      return {
+        ...state,
+        isLoading: false,
+        isError: false,
+        data: action.payload,
+      };
+    case Action.STORIES_FETCH_FAILURE:
+      return {
+        ...state,
+        isLoading: false,
+        isError: true,
+      };
+    case Action.SET_STORIES:
+      return { ...state, data: action.payload };
+    case Action.REMOVE_STORY:
+      return {
+        ...state,
+        data: state.data.filter(
+          (story) => story.objectID !== action.payload.objectID
+        ),
+      };
+    default:
+      throw new Error();
+  }
+};
